entities fixed

+++ src/main/java/ua/roman/flats/repositories/type/TypeConstructionRepository.java
@@ -3,5 +3,5 @@ package ua.roman.flats.repositories.type;
 import org.springframework.data.repository.CrudRepository;
 import ua.roman.flats.domain.type.TypeConstruction;
 
-public interface TypeConstructionRepository extends CrudRepository<TypeConstruction, Long> {
+public interface TypeConstructionRepository extends CrudRepository<TypeConstruction, Integer> {
 }
